rust_lib_app_flutter

Module frb_generated

Re-exports

pub use io::*;

Modules

io 🔒

Structs

FLUTTER_RUST_BRIDGE_HANDLER

IdGenerator 🔒

MoiArc

MoiArcPoolInner

MoiArcPoolValue 🔒

StreamSink

Constants

FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH 🔒

FLUTTER_RUST_BRIDGE_CODEGEN_VERSION 🔒

Traits

CstDecode

MoiArcValue

SseDecode

SseEncode

Functions

decode_rust_opaque_moi 🔒

pde_ffi_dispatcher_primary_impl 🔒

pde_ffi_dispatcher_sync_impl 🔒

transform_result_sse 🔒

wire__crate__api__attach_video_texture_impl 🔒

wire__crate__api__browser_login_begin_impl 🔒

wire__crate__api__browser_login_cancel_impl 🔒

wire__crate__api__browser_login_wait_impl 🔒

wire__crate__api__browser_steam_begin_impl 🔒

wire__crate__api__call_audio_debug_report_impl 🔒

wire__crate__api__call_join_impl 🔒

wire__crate__api__call_leave_impl 🔒

wire__crate__api__call_list_devices_impl 🔒

wire__crate__api__call_list_screen_sources_impl 🔒

wire__crate__api__call_screen_share_start_impl 🔒

wire__crate__api__call_screen_share_stop_impl 🔒

wire__crate__api__call_screen_thumbnail_impl 🔒

wire__crate__api__call_select_device_impl 🔒

wire__crate__api__call_set_camera_enabled_impl 🔒

wire__crate__api__call_set_muted_impl 🔒

wire__crate__api__call_set_screen_audio_enabled_impl 🔒

wire__crate__api__call_set_screen_quality_impl 🔒

wire__crate__api__call_set_stream_volume_impl 🔒

wire__crate__api__call_set_tile_quality_impl 🔒

wire__crate__api__call_set_video_paused_impl 🔒

wire__crate__api__detach_video_texture_impl 🔒

wire__crate__api__edit_message_impl 🔒

wire__crate__api__engine_version_impl 🔒

wire__crate__api__frame_sizes_dto_default_impl 🔒

wire__crate__api__init_app_impl 🔒

wire__crate__api__init_log_impl 🔒

wire__crate__api__init_sentry_impl 🔒

wire__crate__api__link_email_browser_begin_impl 🔒

wire__crate__api__link_email_finish_impl 🔒

wire__crate__api__link_email_start_impl 🔒

wire__crate__api__link_steam_pending_impl 🔒

wire__crate__api__link_steam_via_ticket_impl 🔒

wire__crate__api__list_identities_impl 🔒

wire__crate__api__load_communities_impl 🔒

wire__crate__api__login_finish_impl 🔒

wire__crate__api__login_start_impl 🔒

wire__crate__api__logout_impl 🔒

wire__crate__api__open_space_impl 🔒

wire__crate__api__register_finish_impl 🔒

wire__crate__api__register_start_impl 🔒

wire__crate__api__restore_impl 🔒

wire__crate__api__send_message_impl 🔒

wire__crate__api__sentry_set_user_impl 🔒

wire__crate__api__steam_available_impl 🔒

wire__crate__api__steam_dismiss_impl 🔒

wire__crate__api__steam_login_impl 🔒

wire__crate__api__steam_persona_impl 🔒

wire__crate__api__steam_register_impl 🔒

wire__crate__api__unlink_identity_impl 🔒

wire__crate__api__watch_auth_state_impl 🔒

wire__crate__api__watch_call_state_impl 🔒

wire__crate__api__watch_communities_impl 🔒

wire__crate__api__watch_connection_impl 🔒

wire__crate__api__watch_frame_sizes_impl 🔒

wire__crate__api__watch_levels_impl 🔒

wire__crate__api__watch_messages_impl 🔒

wire__crate__api__watch_roster_impl 🔒

wire__crate__api__watch_stats_impl 🔒

Type Aliases

MoiArcPool

ObjectId 🔒

RustAutoOpaque

Usually this is unneeded, and just write down arbitrary types. However, when you need arbitrary types at places that are not supported yet, use RustOpaqueOpaque<YourArbitraryType>.

RustAutoOpaqueMoi

Please refer to RustAutoOpaque for doc.

RustOpaque

A wrapper to support arbitrary Rust types.

RustOpaqueMoi

Please refer to RustOpaque for doc.